Set up Homebrew shellenv portably
Copied from https://github.com/SteveMarshall/homedir/commit/facf3c80874a7

If Homebrew is installed, we want to set things like PATH to include
Homebrew. On Intel, this mostly happens by magic (because macOS includes
/usr/local/bin, where Homebrew installs things), but on Apple Silicon
it doesn't work by default (because Homebrew on Apple Silicon installs
to /opt/homebrew).

This checks if `brew` is in either of the expected locations, then uses
it to set the shell variables appropriately.

This doesn't just change it because there are rumours of moving it back